Move valid values tests
