+++ tests/test_core_io/test_dtype_validation.py
@@ -1,42 +1,13 @@
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pytest
 
 from geff.core_io import write_arrays
 from geff.metadata._schema import GeffMetadata
-from geff.metadata._valid_values import validate_data_type
 
 if TYPE_CHECKING:
     from geff._typing import PropDictNpArray
-
-
-# -----------------------------------------------------------------------------
-# Unit-tests for `validate_data_type`
-# -----------------------------------------------------------------------------
-@pytest.mark.parametrize(
-    "dtype_in",
-    [
-        "int8",
-        np.int16,
-        np.dtype("uint32"),
-        np.float32,
-        np.dtype("float64"),
-        np.bool_,
-    ],
-)
-def test_validate_data_type_allowed(dtype_in: Any) -> None:
-    """All allowed dtypes should return *True*."""
-    assert validate_data_type(dtype_in) is True
-
-
-@pytest.mark.parametrize(
-    "dtype_in",
-    ["float16", np.float16, "complex64", np.dtype("complex128"), ">f2"],
-)
-def test_validate_data_type_disallowed(dtype_in) -> None:
-    """All disallowed dtypes should return *False*."""
-    assert validate_data_type(dtype_in) is False
 
 
 # -----------------------------------------------------------------------------
